// sistema-hv/src/lib/distribuicao/engine/types.ts
export type ComplexityLevel = 0 | 1 | 2;
export type TemporalLevel = 0 | 1 | 2;
// ...
export interface Process {
  process_id: string;
  theme_id: string;
  collective: boolean;
  complexity_level: ComplexityLevel;
  temporal_level: TemporalLevel;
  directed_executor_id?: string | null;
}

// ---------------------------------------------------------------------------
// Task
// ---------------------------------------------------------------------------

export interface Task {
  task_id: string;
  process_id: string;
  task_type_id: string;
  theme_id: string;
  task_type_points: number;
  theme_multiplier: number;
  task_type_complexity_level: ComplexityLevel;
  task_type_temporal_level: TemporalLevel;
  task_override_complexity_level: ComplexityLevel;
  task_override_temporal_level: ComplexityLevel;
  theme_complexity_level: ComplexityLevel;
  theme_temporal_level: TemporalLevel;
  theme_exclusive_executor_id?: string | null;
  task_type_exclusive_executor_id?: string | null;
  fatal_date: string; // ISO date
  internal_limit_date: string; // ISO date
  input_order: number;
}